Override default $DOCROOT symlink (#140)
By default in the current Tugboat image, `/var/www/html` is a symlink to `/var/www/html.original`, so the `[ -d ]` (is file) shell test prevents the new symlink from being created. Let's just always go ahead and create the symlink, removing the shell test to abort when an existing symlink is found.